--- docs/source/documentation_pages/overview.rst
@@ -1,7 +1,0 @@
-.. _overview-label:
-
-==========================
- Overview of mdlearn
-==========================
-
-**mdlearn** is a Python package for analyzing molecular dynamics simulations machine learning.
