+++ mdlearn/nn/models/aae/__init__.py
@@ -1,1 +1,1 @@
-from .model import AAE  # noqa
+from .model import AAE, ChamferLoss  # noqa
